+++ source/DSCResources/MSFT_SqlDatabase/MSFT_SqlDatabase.psm1
@@ -6,6 +6,17 @@ Import-Module -Name (Join-Path -Path $script:resourceHelperModulePath -ChildPath
 
 $script:localizedData = Get-LocalizedData -ResourceName 'MSFT_SqlDatabase'
 
+$script:supportedCompatibilityLevels = @{
+    8 = @('Version80')
+    9 = @('Version80', 'Version90')
+    10 = @('Version80', 'Version90', 'Version100')
+    11 = @('Version90', 'Version100', 'Version110')
+    12 = @('Version100', 'Version110', 'Version120')
+    13 = @('Version100', 'Version110', 'Version120', 'Version130')
+    14 = @('Version100', 'Version110', 'Version120', 'Version130', 'Version140')
+    15 = @('Version100', 'Version110', 'Version120', 'Version130', 'Version140', 'Version150')
+}
+
 <#
     .SYNOPSIS
     This function gets the sql database.
@@ -26,6 +37,10 @@ $script:localizedData = Get-LocalizedData -ResourceName 'MSFT_SqlDatabase'
     .PARAMETER Collation
     The name of the SQL collation to use for the new database.
     Defaults to server collation.
+
+    .PARAMETER CompatibilityLevel
+    The version of the SQL compatibility level to use for the new database.
+    Defaults to server version.
 #>
 
 function Get-TargetResource
@@ -58,7 +73,12 @@ function Get-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $Collation
+        $Collation,
+
+        [Parameter()]
+        [ValidateSet('Version80', 'Version90', 'Version100', 'Version110', 'Version120', 'Version130', 'Version140', 'Version150')]
+        [System.String]
+        $CompatibilityLevel
     )
 
     Write-Verbose -Message (
@@ -75,9 +95,10 @@ function Get-TargetResource
         {
             $Ensure = 'Present'
             $sqlDatabaseCollation = $sqlDatabaseObject.Collation
+            $sqlDatabaseCompatibilityLevel = $sqlDatabaseObject.CompatibilityLevel
 
             Write-Verbose -Message (
-                $script:localizedData.DatabasePresent -f $Name, $sqlDatabaseCollation
+                $script:localizedData.DatabasePresent -f $Name, $sqlDatabaseCollation, $sqlDatabaseCompatibilityLevel
             )
         }
         else
@@ -91,11 +112,12 @@ function Get-TargetResource
     }
 
     $returnValue = @{
-        Name         = $Name
-        Ensure       = $Ensure
-        ServerName   = $ServerName
-        InstanceName = $InstanceName
-        Collation    = $sqlDatabaseCollation
+        Name               = $Name
+        Ensure             = $Ensure
+        ServerName         = $ServerName
+        InstanceName       = $InstanceName
+        Collation          = $sqlDatabaseCollation
+        CompatibilityLevel = $sqlDatabaseCompatibilityLevel
     }
 
     $returnValue
@@ -121,6 +143,10 @@ function Get-TargetResource
     .PARAMETER Collation
     The name of the SQL collation to use for the new database.
     Defaults to server collation.
+
+    .PARAMETER CompatibilityLevel
+    The version of the SQL compatibility level to use for the new database.
+    Defaults to server version.
 #>
 function Set-TargetResource
 {
@@ -151,7 +177,12 @@ function Set-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $Collation
+        $Collation,
+
+        [Parameter()]
+        [ValidateSet('Version80', 'Version90', 'Version100', 'Version110', 'Version120', 'Version130', 'Version140', 'Version150')]
+        [System.String]
+        $CompatibilityLevel
     )
 
     $sqlServerObject = Connect-SQL -ServerName $ServerName -InstanceName $InstanceName
@@ -169,6 +200,16 @@ function Set-TargetResource
                 New-ObjectNotFoundException -Message $errorMessage
             }
 
+            if (-not $PSBoundParameters.ContainsKey('CompatibilityLevel'))
+            {
+                $CompatibilityLevel = $supportedCompatibilityLevels.$($sqlServerObject.VersionMajor) | Select-Object -Last 1
+            }
+            elseif ($CompatibilityLevel -notin $supportedCompatibilityLevels.$($sqlServerObject.VersionMajor))
+            {
+                $errorMessage = $script:localizedData.InvalidCompatibilityLevel -f $CompatibilityLevel, $InstanceName
+                New-ObjectNotFoundException -Message $errorMessage
+            }
+
             $sqlDatabaseObject = $sqlServerObject.Databases[$Name]
             if ($sqlDatabaseObject)
             {
@@ -179,10 +220,11 @@ function Set-TargetResource
                 try
                 {
                     Write-Verbose -Message (
-                        $script:localizedData.UpdatingCollation -f $Collation
+                        $script:localizedData.UpdatingDatabase -f $Collation, $CompatibilityLevel
                     )
 
                     $sqlDatabaseObject.Collation = $Collation
+                    $sqlDatabaseObject.CompatibilityLevel = $CompatibilityLevel
                     $sqlDatabaseObject.Alter()
                 }
                 catch
@@ -203,6 +245,7 @@ function Set-TargetResource
                         )
 
                         $sqlDatabaseObjectToCreate.Collation = $Collation
+                        $sqlDatabaseObjectToCreate.CompatibilityLevel = $CompatibilityLevel
                         $sqlDatabaseObjectToCreate.Create()
                     }
                 }
@@ -256,6 +299,10 @@ function Set-TargetResource
     .PARAMETER Collation
     The name of the SQL collation to use for the new database.
     Defaults to server collation.
+
+    .PARAMETER CompatibilityLevel
+    The version of the SQL compatibility level to use for the new database.
+    Defaults to server version.
 #>
 function Test-TargetResource
 {
@@ -287,7 +334,12 @@ function Test-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $Collation
+        $Collation,
+
+        [Parameter()]
+        [ValidateSet('Version80', 'Version90', 'Version100', 'Version110', 'Version120', 'Version130', 'Version140', 'Version150')]
+        [System.String]
+        $CompatibilityLevel
     )
 
     Write-Verbose -Message (
@@ -321,13 +373,25 @@ function Test-TargetResource
 
                 $isDatabaseInDesiredState = $false
             }
-            elseif ($PSBoundParameters.ContainsKey('Collation') -and $getTargetResourceResult.Collation -ne $Collation)
+            else
             {
-                Write-Verbose -Message (
-                    $script:localizedData.CollationWrong -f $Name, $getTargetResourceResult.Collation, $Collation
-                )
+                if ($PSBoundParameters.ContainsKey('Collation') -and $getTargetResourceResult.Collation -ne $Collation)
+                {
+                    Write-Verbose -Message (
+                        $script:localizedData.CollationWrong -f $Name, $getTargetResourceResult.Collation, $Collation
+                    )
 
-                $isDatabaseInDesiredState = $false
+                    $isDatabaseInDesiredState = $false
+                }
+
+                if ($PSBoundParameters.ContainsKey('CompatibilityLevel') -and $getTargetResourceResult.CompatibilityLevel -ne $CompatibilityLevel)
+                {
+                    Write-Verbose -Message (
+                        $script:localizedData.CompatibilityLevelWrong -f $Name, $getTargetResourceResult.CompatibilityLevel, $CompatibilityLevel
+                    )
+
+                    $isDatabaseInDesiredState = $false
+                }
             }
         }
     }